chore: move TSGConfig into types
