+++ tsg/__init__.py
@@ -1,9 +1,8 @@
 """
 isort: skip_file
 """
-from .types import Cell
+from .types import Cell, TSGConfig
 from .base_tsg import BaseTSG
-from .tsg_config import TSGConfig
 from .base_manager import BaseManager
 from .stuff import Stuff
 from .stuff_manager import StuffManager
