+++ tsg/types.py
@@ -12,3 +12,21 @@ class Cell:
 
     row: int
     col: int
+
+
+@dataclass
+class TSGConfig:
+    """
+    Config details
+    """
+
+    update_period: float
+    tick_speed: int
+    resolution_w: int
+    resolution_h: int
+    world_width: int
+    world_height: int
+    stuff_chance: int
+    gack_chance: int
+    gack_size: int
+    thing_chance: int
